backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
kafka-python==2.0.2
aiosmtplib==3.0.1
email-validator==2.1.0
structlog==23.2.0
prometheus-client==0.19.0
requests==2.31.0
